Move SequentialGreedyRecommender to subpackage
