+++ baybe/recommenders/__init__.py
@@ -1,9 +1,7 @@
 """BayBE recommenders."""
 
-from baybe.recommenders.bayesian import (
-    NaiveHybridRecommender,
-    SequentialGreedyRecommender,
-)
+from baybe.recommenders.bayesian import NaiveHybridRecommender
+from baybe.recommenders.bayesian.sequential_greedy import SequentialGreedyRecommender
 from baybe.recommenders.nonpredictive.clustering import (
     GaussianMixtureClusteringRecommender,
     KMeansClusteringRecommender,
